fix: Jwt__Key doc, hold cleanup tenant filter, listing delete FK, Razorpay amount overflow
- docs: add required Jwt__Key app setting to startup-diagnostics (fixes 500.30 when missing)
- HoldCleanupHostedService: run cleanup per tenant to avoid tenant mismatch on SaveChanges
- AppDbContext: allow background save when all modified tenant entities share one tenant
- ListingsController.Delete: remove ListingDailyRate/ListingDailyInventory before listing (fix FK conflict)
- RazorpayPaymentService: use long for order amount in paise to avoid Int32 overflow

+++ Atlas.Api/Services/RazorpayPaymentService.cs
@@ -110,9 +110,10 @@ namespace Atlas.Api.Services
                 booking.QuoteTokenNonce = breakdown.QuoteTokenNonce;
                 booking.QuoteExpiresAtUtc = breakdown.QuoteExpiresAtUtc;
 
+                // Razorpay expects amount in smallest currency unit (paise); use long to avoid Int32 overflow for large amounts.
                 var orderRequest = new
                 {
-                    amount = (int)(breakdown.FinalAmount * 100),
+                    amount = (long)(breakdown.FinalAmount * 100),
                     currency = request.Currency,
                     receipt = $"booking_{booking.Id}",
                     payment_capture = 1
